// abertura/src/types/Client.ts
export interface Client {
  nome: string;
  cpf: string;
  rg: string;
  nascimento: string;
  cepR: string;
  enderecoR: string;
  numeroR: string;
  bairroR: string;
  cidadeR: string;
  estadoR: string;
  email: string;
  tel: string;
  tipo: string;
  cnpj: string;
  ie: string;
  razao: string;
  fantasia: string;
  telC: string;
  cepC: string;
  enderecoC: string;
  numeroC: string;
  bairroC: string;
  cidadeC: string;
  estadoC: string;
  setor: string;
  freezer: string;
}